Smartco/Webapps/api.smartco/requirements.txt
bottle
pymysql
requests
pyjwt
phonenumbers
stripe
elasticsearch
uuid
web3
eth_account
langdetect
unidecode
pdfrw
